some fixes and cleanup
