--- x_plane/gui.h
@@ -1,5 +1,0 @@
-#pragma once
-
-void gui_toggle_led( int id , int direction );
-void gui_lcd_put  ( int x , int y , char ch );
-void gui_lcd_clear( );
