+++ x_plane/target.h
@@ -71,10 +71,9 @@
 #define BARO
 #define USE_BARO_BMP280
 
-#define MAG
-#define USE_MAG_HMC5883
-
-#define MAG_HMC5883_ALIGN CW180_DEG
+//#define MAG
+//#define USE_MAG_HMC5883
+//#define MAG_HMC5883_ALIGN CW180_DEG
 
 #define SONAR
 #define BEEPER
@@ -108,7 +107,7 @@
 #define USE_CLI
 
 #define ESC_COUNT 16
-
+#define GTUNE
