+++ x_plane/xplane.cpp
@@ -1,51 +1,54 @@
 #include <qudpsocket.h>
 #include <qthread.h>
+#include "main.h"
 
-extern "C"{
-	#include "xplane.h"
-}
 
 
 
 
 namespace{
 
-	float      acc_data[3];
-	float      gyro_prev[3];
-	float      gyro_data[3];
-	float      bar_pres;
-	float      bar_temp;
-	int        motor_data[8];
-	int        servo_data[8];
+	float        acc_data[3];
+	float        gyro_prev[3];
+	float        gyro_data[3];
+	float        bar_pres;
+	float        bar_temp;
+	int          motor_data[8];
+	int          servo_data[8];
+	QUdpSocket*  send_sock;
+	QHostAddress send_host;
+	int          send_port;
+	char         send_buf[8192];
+	int          send_len;
+	bool         terminate_flag;
+
+
+	void begin( const char* type ){
+		send_len = strlen(type) + 1;
+		memcpy( send_buf , type , send_len );
+	}
+
+
+	void put( double v ){
+		*(float*)(send_buf+send_len) = v;
+		send_len += 4;
+	}
+
+
+	void put( int v ){
+		*(int*)(send_buf+send_len) = v;
+		send_len += 4;
+	}
+
+
+	void end( ){
+		send_sock->writeDatagram( send_buf , send_len , send_host , send_port );
+	}
+
+
 
 	class XplaneThread : public QThread{
 	public:
-		QUdpSocket*  send_sock;
-		QHostAddress send_host;
-		int          send_port;
-		char         send_buf[8192];
-		int          send_len;
-
-		void begin( const char* type ){
-			send_len = strlen(type) + 1;
-			memcpy( send_buf , type , send_len );
-		}
-
-
-		void put( double v ){
-			*(float*)(send_buf+send_len) = v;
-			send_len += 4;
-		}
-
-		void put( int v ){
-			*(int*)(send_buf+send_len) = v;
-			send_len += 4;
-		}
-
-		void end( ){
-			send_sock->writeDatagram( send_buf , send_len , send_host , send_port );
-		}
-
 
 		void run( ) override{
 			QUdpSocket sock;
@@ -56,7 +59,7 @@ namespace{
 			send_port = 49000;
 			send_len  = 0;
 
-			for(;;){
+			while( !terminate_flag ){
 				sock.waitForReadyRead(5);
 
 				while( sock.hasPendingDatagrams() ){
@@ -139,16 +142,17 @@ namespace{
 
 				begin("DATA");
 
-				put( 8 );
-				for( int c=0 ; c < 8 ; c++ ){
-					put( (servo_data[c] - 1000) / 1000.0 );
-				}
+				//put( 8 );
+				//for( int c=0 ; c < 8 ; c++ ){
+				//	put( (servo_data[c] - 1000) / 1000.0 );
+				//}
 
 				put( 25 );
 				for( int c=0 ; c < 8 ; c++ ){
 					put( (motor_data[c] - 1000) / 1000.0 );
 				}
 				end();
+
 			}
 		}
 	};
@@ -157,11 +161,13 @@ namespace{
 }
 
 
+
+
+
 bool xplane_acc_read( int16_t* data ){
 	data[0] = acc_data[0] * 4096;
 	data[1] = acc_data[1] * 4096;
 	data[2] = acc_data[2] * 4096;
-	//printf( "%d %d %d\n" , data[0] , data[1] , data[2] );
 	return true;
 }
 
@@ -172,9 +178,6 @@ bool xplane_gyro_read( int16_t* data ){
 	data[0] = gyro_data[0] * 180 / M_PI * 16.4 ;
 	data[1] = gyro_data[1] * 180 / M_PI * 16.4;
 	data[2] = gyro_data[2] * 180 / M_PI * 16.4;
-
-	//printf( "%f %d\n" , gyro_data[2] , data[2] );
-
 	return true;
 }
 
@@ -186,7 +189,7 @@ bool xplane_gyro_temp( int16_t* data ){
 
 
 void xplane_baro_read( int32_t* pressure , int32_t* temperature ){
-	*pressure    = bar_pres * 33.86389;
+	*pressure    = bar_pres * 3386.389;
 	*temperature = bar_temp * 10;
 }
 
@@ -209,8 +212,15 @@ void xplane_write_servo( int index , int value ){
 }
 
 
-void xplane_init( ){
+void xplane_start( ){
+	terminate_flag = false;
 	thread.start();
+}
+
+
+void xplane_stop( ){
+	terminate_flag = true;
+	thread.wait();
 }
